Schemaをサービス化
