+++ Controller/ApiController.php
@@ -14,31 +14,14 @@
 namespace Plugin\Api\Controller;
 
 use Eccube\Controller\AbstractController;
-use Eccube\Entity\Customer;
-use Eccube\Entity\Order;
-use Eccube\Entity\Product;
-use Eccube\Form\Type\Admin\SearchCustomerType;
-use Eccube\Form\Type\Admin\SearchOrderType;
-use Eccube\Form\Type\Admin\SearchProductType;
-use Eccube\Repository\CustomerRepository;
-use Eccube\Repository\OrderRepository;
-use Eccube\Repository\ProductRepository;
 use GraphQL\Error\DebugFlag;
 use GraphQL\GraphQL;
-use GraphQL\Type\Definition\ObjectType;
-use GraphQL\Type\Definition\Type;
-use GraphQL\Type\Schema;
-use Knp\Component\Pager\Paginator;
-use Plugin\Api\GraphQL\Error\FormInvalidException;
-use Plugin\Api\GraphQL\Type\ConnectionType;
+use Plugin\Api\GraphQL\Schema;
 use Plugin\Api\GraphQL\Types;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
-use Symfony\Component\Form\Extension\Core\Type\IntegerType;
-use Symfony\Component\Form\FormInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\KernelInterface;
 use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Component\Validator\Constraints as Assert;
 
 class ApiController extends AbstractController
 {
@@ -46,41 +29,24 @@ class ApiController extends AbstractController
      * @var Types
      */
     private $types;
-    /**
-     * @var ProductRepository
-     */
-    private $productRepository;
-    /**
-     * @var OrderRepository
-     */
-    private $orderRepository;
-    /**
-     * @var CustomerRepository
-     */
-    private $customerRepository;
+
     /**
      * @var KernelInterface
      */
     private $kernel;
     /**
-     * @var Paginator
+     * @var Schema
      */
-    private $paginator;
+    private $schema;
 
     public function __construct(
         Types $types,
-        ProductRepository $productRepository,
-        OrderRepository $orderRepository,
-        CustomerRepository $customerRepository,
         KernelInterface $kernel,
-        Paginator $paginator
+        Schema $schema
     ) {
         $this->types = $types;
-        $this->productRepository = $productRepository;
-        $this->orderRepository = $orderRepository;
-        $this->customerRepository = $customerRepository;
         $this->kernel = $kernel;
-        $this->paginator = $paginator;
+        $this->schema = $schema;
     }
 
     /**
@@ -90,7 +56,7 @@ class ApiController extends AbstractController
     public function index(Request $request)
     {
         $body = json_decode($request->getContent(), true);
-        $schema = $this->getSchema();
+        $schema = $this->schema;
         $query = $body['query'];
         $variableValues = isset($body['variables']) ? $body['variables'] : null;
         $result = GraphQL::executeQuery($schema, $query, null, null, $variableValues);
@@ -101,100 +67,5 @@ class ApiController extends AbstractController
         }
 
         return $this->json($result);
-    }
-
-    /**
-     * @return Schema
-     */
-    private function getSchema()
-    {
-        return new Schema([
-            'query' => new ObjectType([
-                'name' => 'Query',
-                'fields' => [
-                    'products' => $this->createQuery(Product::class, SearchProductType::class, function ($searchData) {
-                        return $this->productRepository->getQueryBuilderBySearchDataForAdmin($searchData);
-                    }),
-                    'orders' => $this->createQuery(Order::class, SearchOrderType::class, function ($searchData) {
-                        return $this->orderRepository->getQueryBuilderBySearchDataForAdmin($searchData);
-                    }),
-                    'customers' => $this->createQuery(Customer::class, SearchCustomerType::class, function ($searchData) {
-                        return $this->customerRepository->getQueryBuilderBySearchData($searchData);
-                    }),
-                ],
-                'typeLoader' => function ($name) {
-                    return $this->types->get($name);
-                },
-            ]),
-        ]);
-    }
-
-    private function createQuery($entityClass, $searchFormType, $resolver)
-    {
-        $builder = $this->formFactory->createBuilder($searchFormType, null, ['csrf_protection' => false]);
-
-        // paging のためのフォームを追加
-        $builder->add('page', IntegerType::class, [
-            'label' => 'api.args.page.description',
-            'required' => false,
-            'data' => 1,
-            'constraints' => [
-                new Assert\Regex([
-                    'pattern' => "/^\d+$/u",
-                    'message' => 'form_error.numeric_only',
-                ]),
-            ],
-        ])->add('limit', IntegerType::class, [
-            'label' => 'api.args.limit.description',
-            'required' => false,
-            'data' => $this->eccubeConfig->get('eccube_default_page_count'),
-            'constraints' => [
-                new Assert\Regex([
-                    'pattern' => "/^\d+$/u",
-                    'message' => 'form_error.numeric_only',
-                ]),
-            ],
-        ]);
-
-        $args = array_reduce($builder->getForm()->all(), function ($acc, $form) {
-            /* @var FormInterface $form */
-            $formConfig = $form->getConfig();
-            $type = $formConfig->getType()->getInnerType() instanceof IntegerType ? Type::int() : Type::string();
-            if ($formConfig->getOption('multiple')) {
-                $type = Type::listOf($type);
-            }
-            if ($formConfig->getOption('required') && !$formConfig->getOption('multiple')) {
-                $type = Type::nonNull($type);
-            }
-            $acc[$form->getName()] = [
-                'type' => $type,
-                'defaultValue' => $form->getViewData(),
-                'description' => $formConfig->getOption('label') ? trans($formConfig->getOption('label')) : null,
-            ];
-
-            return $acc;
-        }, []);
-
-        return [
-            'type' => new ConnectionType($entityClass, $this->types),
-            'args' => $args,
-            'resolve' => function ($root, $args) use ($builder, $resolver) {
-                $form = $builder->getForm();
-                $form->submit($args);
-
-                if (!$form->isValid()) {
-                    $message = '';
-                    foreach ($form->getErrors(true) as $error) {
-                        $message .= sprintf('%s: %s;', $error->getOrigin()->getName(), $error->getMessage());
-                    }
-
-                    throw new FormInvalidException($message);
-                }
-
-                $data = $form->getData();
-
-                return $this->paginator->paginate($resolver($data), $args['page'], $args['limit']);
-            },
-        ];
     }
 }
